refactor amendment page components
